BigInteger in Tweet

+++ twitoff2/dbmodel.py
@@ -14,10 +14,10 @@ class User(DB.Model):
         return '<User %r>' % self.username
 
 class Tweet(DB.Model):
-    Id = DB.Column(DB.Integer, primary_key=True)
+    Id = DB.Column(DB.BigInteger, primary_key=True)
     tweet = DB.Column(DB.String(280), nullable=False)
     embedding = DB.Column(DB.PickleType, nullable = False)
-    user_id = DB.Column(DB.Integer, DB.ForeignKey('user.Id'), nullable=False)
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.Id'), nullable=False)
     user = DB.relationship('User', backref=DB.backref('tweet', lazy=True))
 
     def __repr__(self):
